Expose debug for inline helpers
